requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
drf-yasg==1.21.7
mysqlclient==2.2.4
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.0
toml==0.10.2
tzdata==2022.6
